+++ crates/oxc_mangler/src/lib.rs
@@ -194,7 +194,7 @@ impl Mangler {
             // rename the variables
             for (symbol_to_rename, new_name) in symbols_to_rename_with_new_names {
                 for &symbol_id in &symbol_to_rename.symbol_ids {
-                    symbol_table.set_name(symbol_id, new_name.clone());
+                    symbol_table.set_name(symbol_id, new_name.as_str());
                 }
             }
         }