+++ crates/oxc_semantic/src/scope.rs
@@ -1,10 +1,9 @@
-use std::mem;
+use std::{fmt, mem};
 
-use rustc_hash::{FxBuildHasher, FxHashMap};
+use rustc_hash::FxBuildHasher;
 
-use oxc_allocator::Allocator;
-use oxc_index::IndexVec;
-use oxc_span::CompactStr;
+use oxc_allocator::{Allocator, Vec as ArenaVec};
+use oxc_index::{Idx, IndexVec};
 use oxc_syntax::{
     node::NodeId,
     reference::ReferenceId,
@@ -13,7 +12,8 @@ use oxc_syntax::{
 };
 
 pub(crate) type Bindings<'a> = hashbrown::HashMap<&'a str, SymbolId, FxBuildHasher, &'a Allocator>;
-pub type UnresolvedReferences = FxHashMap<CompactStr, Vec<ReferenceId>>;
+pub type UnresolvedReferences<'a> =
+    hashbrown::HashMap<&'a str, ArenaVec<'a, ReferenceId>, FxBuildHasher, &'a Allocator>;
 
 /// Scope Tree
 ///
@@ -29,9 +29,6 @@ pub struct ScopeTree {
     /// Maps a scope to the parent scope it belongs in.
     parent_ids: IndexVec<ScopeId, Option<ScopeId>>,
 
-    /// Maps a scope to direct children scopes.
-    child_ids: IndexVec<ScopeId, Vec<ScopeId>>,
-
     /// Runtime flag for constructing child_ids.
     pub(crate) build_child_ids: bool,
 
@@ -40,22 +37,29 @@ pub struct ScopeTree {
 
     flags: IndexVec<ScopeId, ScopeFlags>,
 
-    pub(crate) root_unresolved_references: UnresolvedReferences,
-
     pub(crate) cell: ScopeTreeCell,
+}
+
+impl fmt::Debug for ScopeTree {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
+        f.debug_struct("ScopeTree").finish()
+    }
 }
 
 impl Default for ScopeTree {
     fn default() -> Self {
         Self {
             parent_ids: IndexVec::new(),
-            child_ids: IndexVec::new(),
             build_child_ids: false,
             node_ids: IndexVec::new(),
             flags: IndexVec::new(),
-            root_unresolved_references: UnresolvedReferences::default(),
-            cell: ScopeTreeCell::new(Allocator::default(), |_bump| ScopeTreeInner {
+            cell: ScopeTreeCell::new(Allocator::default(), |allocator| ScopeTreeInner {
                 bindings: IndexVec::new(),
+                child_ids: ArenaVec::new_in(allocator),
+                root_unresolved_references: UnresolvedReferences::with_hasher_in(
+                    FxBuildHasher,
+                    allocator,
+                ),
             }),
         }
     }
@@ -74,6 +78,11 @@ pub(crate) struct ScopeTreeInner<'cell> {
     ///
     /// A binding is a mapping from an identifier name to its [`SymbolId`]
     pub(crate) bindings: IndexVec<ScopeId, Bindings<'cell>>,
+
+    /// Maps a scope to direct children scopes.
+    child_ids: ArenaVec<'cell, ArenaVec<'cell, ScopeId>>,
+
+    pub(crate) root_unresolved_references: UnresolvedReferences<'cell>,
 }
 
 impl ScopeTree {
@@ -125,13 +134,28 @@ impl ScopeTree {
 
     #[inline]
     pub fn root_unresolved_references(&self) -> &UnresolvedReferences {
-        &self.root_unresolved_references
+        &self.cell.borrow_dependent().root_unresolved_references
     }
 
     pub fn root_unresolved_references_ids(
         &self,
     ) -> impl Iterator<Item = impl Iterator<Item = ReferenceId> + '_> + '_ {
-        self.root_unresolved_references.values().map(|v| v.iter().copied())
+        self.cell.borrow_dependent().root_unresolved_references.values().map(|v| v.iter().copied())
+    }
+
+    pub(crate) fn set_root_unresolved_references<'a>(
+        &mut self,
+        entries: impl Iterator<Item = (&'a str, Vec<ReferenceId>)>,
+    ) {
+        self.cell.with_dependent_mut(|allocator, inner| {
+            for (k, v) in entries {
+                let k = allocator.alloc_str(k);
+                let v = ArenaVec::from_iter_in(v, allocator);
+                inner.root_unresolved_references.insert(k, v);
+            }
+            // =
+            // .extend_from(entries.map(|(k, v)| (allocator.alloc(k),)))
+        });
     }
 
     /// Delete an unresolved reference.
@@ -147,14 +171,16 @@ impl ScopeTree {
         // but `map.entry` requires an owned key to be provided. Currently we use `CompactStr`s as keys
         // which are not cheap to construct, so this is best we can do at present.
         // TODO: Switch to `Entry` API once we use `&str`s or `Atom`s as keys.
-        let reference_ids = self.root_unresolved_references.get_mut(name).unwrap();
-        if reference_ids.len() == 1 {
-            assert!(reference_ids[0] == reference_id);
-            self.root_unresolved_references.remove(name);
-        } else {
-            let index = reference_ids.iter().position(|&id| id == reference_id).unwrap();
-            reference_ids.swap_remove(index);
-        }
+        self.cell.with_dependent_mut(|_allocator, inner| {
+            let reference_ids = inner.root_unresolved_references.get_mut(name).unwrap();
+            if reference_ids.len() == 1 {
+                assert!(reference_ids[0] == reference_id);
+                inner.root_unresolved_references.remove(name);
+            } else {
+                let index = reference_ids.iter().position(|&id| id == reference_id).unwrap();
+                reference_ids.swap_remove(index);
+            }
+        });
     }
 
     #[inline]
@@ -183,7 +209,9 @@ impl ScopeTree {
         if self.build_child_ids {
             // Set this scope as child of parent scope
             if let Some(parent_id) = parent_id {
-                self.child_ids[parent_id].push(scope_id);
+                self.cell.with_dependent_mut(|_allocator, inner| {
+                    inner.child_ids[parent_id.index()].push(scope_id);
+                });
             }
         }
     }
@@ -194,25 +222,29 @@ impl ScopeTree {
     pub fn change_parent_id(&mut self, scope_id: ScopeId, new_parent_id: Option<ScopeId>) {
         let old_parent_id = mem::replace(&mut self.parent_ids[scope_id], new_parent_id);
         if self.build_child_ids {
-            // Remove this scope from old parent scope
-            if let Some(old_parent_id) = old_parent_id {
-                self.child_ids[old_parent_id].retain(|&child_id| child_id != scope_id);
-            }
-            // And add it to new parent scope
-            if let Some(parent_id) = new_parent_id {
-                self.child_ids[parent_id].push(scope_id);
-            }
+            self.cell.with_dependent_mut(|_allocator, inner| {
+                // Remove this scope from old parent scope
+                if let Some(old_parent_id) = old_parent_id {
+                    inner.child_ids[old_parent_id.index()].retain(|&child_id| child_id != scope_id);
+                }
+                // And add it to new parent scope
+                if let Some(parent_id) = new_parent_id {
+                    inner.child_ids[parent_id.index()].push(scope_id);
+                }
+            });
         }
     }
 
     /// Delete a scope.
     pub fn delete_scope(&mut self, scope_id: ScopeId) {
         if self.build_child_ids {
-            self.child_ids[scope_id].clear();
-            let parent_id = self.parent_ids[scope_id];
-            if let Some(parent_id) = parent_id {
-                self.child_ids[parent_id].retain(|&child_id| child_id != scope_id);
-            }
+            self.cell.with_dependent_mut(|_allocator, inner| {
+                inner.child_ids[scope_id.index()].clear();
+                let parent_id = self.parent_ids[scope_id];
+                if let Some(parent_id) = parent_id {
+                    inner.child_ids[parent_id.index()].retain(|&child_id| child_id != scope_id);
+                }
+            });
         }
     }
 
@@ -222,8 +254,15 @@ impl ScopeTree {
         self.get_binding(self.root_scope_id(), name)
     }
 
-    pub fn add_root_unresolved_reference(&mut self, name: CompactStr, reference_id: ReferenceId) {
-        self.root_unresolved_references.entry(name).or_default().push(reference_id);
+    pub fn add_root_unresolved_reference(&mut self, name: &str, reference_id: ReferenceId) {
+        self.cell.with_dependent_mut(|allocator, inner| {
+            let name = allocator.alloc_str(name);
+            inner
+                .root_unresolved_references
+                .entry(name)
+                .or_insert_with(|| ArenaVec::new_in(allocator))
+                .push(reference_id);
+        });
     }
 
     /// Check if a symbol is declared in a certain scope.
@@ -304,15 +343,18 @@ impl ScopeTree {
     /// Get the child scopes of a scope
     #[inline]
     pub fn get_child_ids(&self, scope_id: ScopeId) -> &[ScopeId] {
-        &self.child_ids[scope_id]
+        &self.cell.borrow_dependent().child_ids[scope_id.index()]
     }
 
     pub fn iter_all_child_ids(&self, scope_id: ScopeId) -> impl Iterator<Item = ScopeId> + '_ {
-        let mut stack = self.child_ids[scope_id].clone();
-        let child_ids: &IndexVec<ScopeId, Vec<ScopeId>> = &self.child_ids;
+        let mut stack = self.cell.borrow_dependent().child_ids[scope_id.index()]
+            .iter()
+            .copied()
+            .collect::<Vec<_>>();
+        let child_ids = &self.cell.borrow_dependent().child_ids;
         std::iter::from_fn(move || {
             if let Some(scope_id) = stack.pop() {
-                if let Some(children) = child_ids.get(scope_id) {
+                if let Some(children) = child_ids.get(scope_id.index()) {
                     stack.extend(children.iter().copied());
                 }
                 Some(scope_id)
@@ -322,10 +364,11 @@ impl ScopeTree {
         })
     }
 
-    /// Get a mutable reference to a scope's children
-    #[inline]
-    pub fn get_child_ids_mut(&mut self, scope_id: ScopeId) -> &mut Vec<ScopeId> {
-        &mut self.child_ids[scope_id]
+    pub fn remove_child_scopes(&mut self, scope_id: ScopeId, child_scope_ids: &[ScopeId]) {
+        self.cell.with_dependent_mut(|_allocator, inner| {
+            inner.child_ids[scope_id.index()]
+                .retain(|scope_id| !child_scope_ids.contains(scope_id));
+        });
     }
 
     /// Create a scope.
@@ -343,10 +386,12 @@ impl ScopeTree {
         });
         self.node_ids.push(node_id);
         if self.build_child_ids {
-            self.child_ids.push(vec![]);
-            if let Some(parent_id) = parent_id {
-                self.child_ids[parent_id].push(scope_id);
-            }
+            self.cell.with_dependent_mut(|allocator, inner| {
+                inner.child_ids.push(ArenaVec::new_in(allocator));
+                if let Some(parent_id) = parent_id {
+                    inner.child_ids[parent_id.index()].push(scope_id);
+                }
+            });
         }
         scope_id
     }
@@ -412,7 +457,9 @@ impl ScopeTree {
         });
         self.node_ids.reserve(additional);
         if self.build_child_ids {
-            self.child_ids.reserve(additional);
+            self.cell.with_dependent_mut(|_allocator, inner| {
+                inner.child_ids.reserve(additional);
+            });
         }
     }
 }